+++ packages/headless/src/playwright.ts
@@ -1,2 +1,3 @@
 export * from "./composables/comboBox/createComboBox.ct";
 export * from "./composables/listbox/createListbox.ct";
+export * from "./composables/menuButton/createMenuButton.ct";