+++ packages/sit-onyx/src/components/OnyxListItem/OnyxListItem.vue
@@ -92,4 +92,3 @@ const props = defineProps<Pick<SelectOption, "color"> & { active?: boolean }>();
   }
 }
 </style>
-../OnyxSelect/types