+++ packages/sit-onyx/src/components/OnyxSelect/OnyxSelect.vue
@@ -51,6 +51,10 @@ const slots = defineSlots<{
    * to e.g. show a button to load more options instead of lazy loading on scroll.
    */
   optionsEnd?(): unknown;
+  /**
+   * Optional slot to override the option content.
+   */
+  option?(props: SelectOption<TValue>): unknown;
 }>();
 
 const { t } = injectI18n();
@@ -258,6 +262,7 @@ const selectInputProps = computed(() => {
       ref="selectInput"
       v-bind="selectInputProps"
       :show-focus="isExpanded"
+      :autofocus="props.autofocus"
       @click="onToggle"
     />
 
@@ -337,7 +342,9 @@ const selectInputProps = computed(() => {
                 :color="option.color"
                 :density="props.density"
               >
-                {{ option.label }}
+                <slot name="option" v-bind="option">
+                  {{ option.label }}
+                </slot>
               </OnyxSelectOption>
             </ul>
           </template>