+++ packages/sit-onyx/src/components/OnyxSelectInput/OnyxSelectInput.vue
@@ -2,7 +2,6 @@
 import chevronDownUp from "@sit-onyx/icons/chevron-down-up.svg?raw";
 import { computed, ref } from "vue";
 import { useDensity } from "../../composables/density";
-import { useRequired } from "../../composables/required";
 import { injectI18n } from "../../i18n";
 import type { SelectOptionValue } from "../../types";
 import { useRootAttrs } from "../../utils/attrs";
@@ -12,6 +11,7 @@ import OnyxLoadingIndicator from "../OnyxLoadingIndicator/OnyxLoadingIndicator.v
 import OnyxSkeleton from "../OnyxSkeleton/OnyxSkeleton.vue";
 import OnyxTooltip from "../OnyxTooltip/OnyxTooltip.vue";
 import type { OnyxSelectInputProps } from "./types";
+import OnyxFormElement from "../OnyxFormElement/OnyxFormElement.vue";
 
 defineOptions({ inheritAttrs: false });
 const { rootAttrs, restAttrs } = useRootAttrs();
@@ -60,7 +60,6 @@ const selectionText = computed<string>(() => {
   return props.selection?.label ?? "";
 });
 
-const { requiredMarkerClass, requiredTypeClass } = useRequired(props);
 const { densityClass } = useDensity(props);
 
 const input = ref<HTMLInputElement>();
@@ -81,20 +80,12 @@ defineExpose({ focus: () => input.value?.focus() });
     v-else
     :class="[
       'onyx-select-input',
-      requiredTypeClass,
       densityClass,
       props.readonly ? 'onyx-select-input--readonly' : 'onyx-select-input--editable',
     ]"
     v-bind="rootAttrs"
   >
-    <label>
-      <div
-        v-if="!props.hideLabel"
-        :class="['onyx-select-input__label', 'onyx-text--small', requiredMarkerClass]"
-      >
-        <div class="onyx-truncation-ellipsis">{{ props.label }}</div>
-      </div>
-
+    <OnyxFormElement v-bind="props">
       <div class="onyx-select-input__wrapper">
         <OnyxLoadingIndicator
           v-if="props.loading"
@@ -105,8 +96,8 @@ defineExpose({ focus: () => input.value?.focus() });
         <input
           ref="input"
           :class="{
-            'onyx-select-input__input': true,
-            'onyx-select-input__input--show-focus': props.showFocus,
+            'onyx-select-input__native': true,
+            'onyx-select-input__native--show-focus': props.showFocus,
             'onyx-truncation-ellipsis': true,
           }"
           v-bind="restAttrs"
@@ -118,6 +109,7 @@ defineExpose({ focus: () => input.value?.focus() });
           :aria-label="props.hideLabel ? props.label : undefined"
           :title="props.hideLabel ? props.label : undefined"
           :value="selectionText"
+          :autofocus="props.autofocus"
           @click="emit('click')"
         />
 
@@ -144,20 +136,14 @@ defineExpose({ focus: () => input.value?.focus() });
           <OnyxIcon :icon="chevronDownUp" />
         </button>
       </div>
-    </label>
-
-    <div
-      v-if="props.message"
-      class="onyx-select-input__footer onyx-text--small onyx-truncation-ellipsis"
-    >
-      {{ props.message }}
-    </div>
+    </OnyxFormElement>
   </div>
 </template>
 
 <style lang="scss">
 @use "../../styles/mixins/density.scss";
 @use "../../styles/mixins/layers.scss";
+@use "../../styles/mixins/input.scss";
 
 .onyx-select-input,
 .onyx-select-input-skeleton {
@@ -176,26 +162,16 @@ defineExpose({ focus: () => input.value?.focus() });
 
 .onyx-select-input {
   @include layers.component() {
-    $line-height: 1.5rem;
+    @include input.define-shared-styles(
+      $base-selector: ".onyx-select-input",
+      $vertical-padding: var(--onyx-select-input-padding-vertical)
+    );
 
-    --border-color: var(--onyx-color-base-neutral-300);
     --selection-color: var(--onyx-color-base-neutral-200);
-
     font-family: var(--onyx-font-family);
     display: flex;
     flex-direction: column;
     gap: var(--onyx-spacing-5xs);
-
-    &__label {
-      display: flex;
-      margin-bottom: var(--onyx-spacing-5xs);
-      color: var(--onyx-color-text-icons-neutral-medium);
-
-      // optional marker should be displayed at the very end of the label
-      &.onyx-optional-marker {
-        justify-content: space-between;
-      }
-    }
 
     &__button {
       all: initial;
@@ -204,50 +180,6 @@ defineExpose({ focus: () => input.value?.focus() });
 
       &:enabled {
         cursor: pointer;
-      }
-    }
-
-    &__wrapper {
-      border-radius: var(--onyx-radius-sm);
-      border: var(--onyx-1px-in-rem) solid var(--border-color);
-      background: var(--onyx-color-base-background-blank);
-      color: var(--onyx-color-text-icons-neutral-intense);
-
-      display: flex;
-      align-items: center;
-      gap: var(--onyx-spacing-2xs);
-
-      font-size: 1rem;
-      line-height: $line-height;
-
-      box-sizing: border-box;
-
-      padding: var(--onyx-select-input-padding-vertical) var(--onyx-spacing-sm);
-      height: calc($line-height + 2 * var(--onyx-select-input-padding-vertical));
-    }
-
-    &__input {
-      // reset native input styles so they are inherited from the parent
-      border: none;
-      border-radius: inherit;
-      background-color: transparent;
-      color: inherit;
-      width: 100%;
-      outline: none;
-      font-family: inherit;
-      font-size: inherit;
-      line-height: inherit;
-      padding: 0;
-      cursor: inherit;
-
-      &::placeholder {
-        color: var(--onyx-color-text-icons-neutral-soft);
-        opacity: 1;
-        font-weight: 400;
-      }
-
-      &::selection {
-        background: var(--selection-color);
       }
     }
 
@@ -260,17 +192,19 @@ defineExpose({ focus: () => input.value?.focus() });
       color: var(--onyx-color-text-icons-primary-intense);
     }
 
-    &__footer {
-      width: 100%;
-      color: var(--onyx-color-text-icons-neutral-soft);
-    }
-
+    /** The internal input is always "readonly" because typing is not allowed.
+     * That's why we need to rely on modifier classes instead of using pseudo classes
+     */
     &--editable {
-      .onyx-select-input__wrapper:has(.onyx-select-input__input:enabled) {
+      .onyx-select-input__wrapper:has(.onyx-select-input__native:enabled) {
         cursor: pointer;
+        .onyx-select-input__native {
+          cursor: pointer;
+        }
+
         // default hover
         &:hover {
-          --border-color: var(--onyx-color-base-primary-400);
+          @include input.define-enabled-hover();
           .onyx-select-input__button {
             color: var(--onyx-color-text-icons-primary-medium);
           }
@@ -278,13 +212,12 @@ defineExpose({ focus: () => input.value?.focus() });
       }
       // default focus
       &:has(
-          .onyx-select-input__input:enabled:focus,
-          .onyx-select-input__input--show-focus:enabled
+          .onyx-select-input__native:enabled:focus,
+          .onyx-select-input__native--show-focus:enabled
         ) {
         .onyx-select-input {
           &__wrapper {
-            --border-color: var(--onyx-color-base-primary-500);
-            outline: var(--onyx-spacing-4xs) solid var(--onyx-color-base-primary-200);
+            @include input.define-enabled-focus();
           }
 
           &__button {
@@ -296,20 +229,17 @@ defineExpose({ focus: () => input.value?.focus() });
 
     // readonly focus
     &--readonly:has(
-        .onyx-select-input__input:enabled:focus,
-        .onyx-select-input__input--show-focus:enabled
+        .onyx-select-input__native:enabled:focus,
+        .onyx-select-input__native--show-focus:enabled
       )
       .onyx-select-input__wrapper {
       outline: var(--onyx-spacing-4xs) solid var(--onyx-color-base-neutral-200);
+      --border-color: var(--onyx-color-base-neutral-400);
     }
 
-    &:has(&__input:disabled),
+    &:has(&__native:disabled),
     &--readonly {
       .onyx-select-input {
-        &__label {
-          color: var(--onyx-color-text-icons-neutral-soft);
-        }
-
         &__wrapper {
           background-color: var(--onyx-color-base-background-tinted);
           color: var(--onyx-color-text-icons-neutral-soft);
@@ -319,23 +249,19 @@ defineExpose({ focus: () => input.value?.focus() });
     }
 
     &--readonly {
+      .onyx-select-input__native:enabled {
+        cursor: initial;
+      }
+
       .onyx-select-input__wrapper:hover {
         --border-color: var(--onyx-color-base-neutral-400);
       }
     }
 
     &-skeleton {
-      display: flex;
-      flex-direction: column;
-      gap: var(--onyx-spacing-5xs);
-      &__label {
-        width: var(--onyx-spacing-3xl);
-        height: 1.25rem;
-      }
-      &__input {
-        width: 17rem;
-        height: calc($line-height + 2 * var(--onyx-select-input-padding-vertical));
-      }
+      @include input.define-skeleton-styles(
+        $height: calc(1lh + 2 * var(--onyx-select-input-padding-vertical))
+      );
     }
   }
 }