+++ packages/chartjs-plugin/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @sit-onyx/chartjs-plugin
 
+## 1.0.0-alpha.97
+
+### Patch Changes
+
+- Updated dependencies [08b434b]
+  - sit-onyx@1.0.0-alpha.141
+
+## 1.0.0-alpha.96
+
+### Patch Changes
+
+- Updated dependencies [d3e9321]
+  - sit-onyx@1.0.0-alpha.140
+
+## 1.0.0-alpha.95
+
+### Patch Changes
+
+- Updated dependencies [107ec36]
+  - sit-onyx@1.0.0-alpha.139
+
+## 1.0.0-alpha.94
+
+### Patch Changes
+
+- Updated dependencies [0863114]
+- Updated dependencies [0863114]
+- Updated dependencies [53b0d50]
+  - sit-onyx@1.0.0-alpha.138
+
 ## 1.0.0-alpha.93
 
 ### Patch Changes