+++ packages/sit-onyx/CHANGELOG.md
@@ -1,5 +1,36 @@
 # sit-onyx
 
+## 1.0.0-alpha.141
+
+### Minor Changes
+
+- 08b434b: feat(OnyxSelectInput): support message/label tooltip
+
+## 1.0.0-alpha.140
+
+### Minor Changes
+
+- d3e9321: Add new OnyxNavButton component
+
+## 1.0.0-alpha.139
+
+### Minor Changes
+
+- 107ec36: feat(OnyxSelect): support `option` slot
+
+  The new slot can be used to show custom content for the select options
+
+## 1.0.0-alpha.138
+
+### Minor Changes
+
+- 0863114: feat: support `autofocus` property for multiple components
+
+  - OnyxButton, OnyxCheckbox, OnyxIconButton, OnyxRadioButton, OnyxSwitch, OnyxSelect
+
+- 0863114: feat: add `OnyxColorSchemeDialog` component
+- 53b0d50: feat(OnyxInput,OnyxTextarea): Unify template & style of header/footer form elements
+
 ## 1.0.0-alpha.137
 
 ### Patch Changes